+++ routes/api/transactions.js
@@ -3,22 +3,9 @@ const { BadRequest } = require("http-errors");
 const router = express.Router();
 
 const { Transaction, User } = require("../../models");
-const {
-  joiTransactionValidation,
-  joiTransactionStatisticValidation,
-  authenticate,
-} = require("../../middlewares");
-const {
-  amountByCategory,
-  amountByTransactionType,
-  countTheBalance,
-} = require("../../helpers/index");
+const { joiTransactionValidation, authenticate } = require("../../middlewares");
 
-/* 
-1. Получение всех транзакций пользователя.
-2. Создание транзакции.
-3. Получение статистики транзакций по месяцу и году. 
-*/
+const { countTheBalance } = require("../../helpers/index");
 
 /*энд-поинт получение всех транзакций пользователя.*/
 router.get("/", authenticate, async (req, res, next) => {
@@ -65,44 +52,12 @@ router.post("/", authenticate, async (req, res, next) => {
       owner: _id,
       balance: transactionBalance,
     });
+
     res.status(201).json(newTransaction);
   } catch (error) {
     if (error.message.includes("validation failed")) {
       error.status = 400;
     }
-    next(error);
-  }
-});
-
-/* энд-поинт на получение подробной статистики 
-за месяц и год по транзакциям пользователя*/
-router.get("/statistic", authenticate, async (req, res, next) => {
-  try {
-    const { error } = joiTransactionStatisticValidation.validate(req.body);
-    if (error) {
-      throw new BadRequest(error.message);
-    }
-
-    const { _id } = req.user;
-    const { year, month } = req.body;
-
-    const transactions = await Transaction.find({
-      owner: _id,
-      year: year,
-      month: month,
-    });
-
-    const categoriesTransaction = amountByCategory(transactions);
-    const totalIncome = amountByTransactionType(transactions, true);
-    const totalExpence = amountByTransactionType(transactions, false);
-
-    const data = {
-      category: { ...categoriesTransaction },
-      total: { Expence: totalExpence, Income: totalIncome },
-    };
-
-    res.json(data);
-  } catch (error) {
     next(error);
   }
 });
