+++ routes/index.js
@@ -1,7 +1,9 @@
 const usersRouter = require("./api/users.js");
 const transactionsRouter = require("./api/transactions.js");
+const statisticsRouter = require("./api/statistics.js");
 
 module.exports = {
   usersRouter,
   transactionsRouter,
+  statisticsRouter,
 };
